Move Connection preferences to Qt UI file
