+++ lang/1033.json
@@ -247,6 +247,7 @@
         "pause_on_low_disk_space": "Pause when disk space is low",
         "pause_on_low_disk_space_alert": "Torrent paused (low disk space)",
         "enable_geoip": "Enable GeoIP lookups",
-        "enable_automatic_crash_reporting": "Enable automatic crash reporting"
+        "enable_automatic_crash_reporting": "Enable automatic crash reporting",
+        "network_adapter": "Network adapter"
     }
 }
